+++ src/text/wrap.zig
@@ -321,16 +321,8 @@ fn appendGlyphSlice(
 // Tests
 // ---------------------------------------------------------------------------
 
-const test_font_paths = [_][:0]const u8{
-    "/System/Library/Fonts/Helvetica.ttc",
-    "/System/Library/Fonts/Monaco.ttf",
-};
-
 fn loadTestFont(fs: *FontSystem) !FontId {
-    for (test_font_paths) |path| {
-        if (fs.loadFont(path, 0)) |id| return id else |_| {}
-    }
-    return error.SkipZigTest;
+    return font.loadTestFont(fs);
 }
 
 test "wrap hello world forces multiple lines" {